moved the auth facade class into the auth directory
