+++ masonite/auth/__init__.py
@@ -1,2 +1,3 @@
+from .Auth import Auth
 from .Csrf import Csrf
 from .Sign import Sign
