--- masonite/facades/__init__.py
@@ -1,1 +1,0 @@
-from .Auth import Auth
