+++ masonite/snippets/auth/controllers/RegisterController.py
@@ -1,5 +1,5 @@
 ''' A Module Description '''
-from masonite.facades.Auth import Auth
+from masonite.auth import Auth
 from config import auth
 import bcrypt
